Description
Text Mining is cleaning the texts by removing the unwanted or useless data and make the data much more meaningful. 
Text mining involves removing punctuations, numbers, similar words, lower casing all the words, removing the null spaces etc., There are some predefined english words called ‘Stopwords’ which are to be removed to make the data more clean. In addition to the predefined stopwords, new stopwords can be added. The text document or the dataset must be saved as corpus for easy text mining.
Text mining is done in R by using the package called ‘tm’. “Tm_map” is used to perform all kinds of text manipulating operations that are mentioned above. Once all the stops words are removed there are lots of spaces which are of no use are available in the text document. So, now we have to remove those stripwhitespaces. Another technique in Text mining is Stem Document which converts all the words to a singular verb form so that it makes the words much more accurate for further operations.